Centralize dev & test Graph API configs
