+++ .env.test
@@ -1,5 +1,3 @@
-DEV_USER_ID=10000
-
 LOGTO_ENDPOINT_URL=https://u4v5ne.logto.app/
 LOGTO_SIGN_IN_CALLBACK_URL=http://localhost:3000/login
 LOGTO_SIGN_OUT_REDIRECT_URL=http://localhost:3000/login
