+++ docs/docs/design.mdx
@@ -79,7 +79,7 @@ Our free version comes with an in-memory [json-graphql-server]. We can host our 
 
 ```bash
 cd nexusgraph
-yarn start:json-graphql-server-dev
+yarn start:graph-api
 ```
 
 The server will be running at `http://localhost:5000/`.
